requirements.txt
# Core API
fastapi==0.129.0
uvicorn==0.41.0
python-multipart==0.0.22
pydantic==2.12.5

# RAG pipeline
groq==1.0.0
google-genai==1.64.0
chromadb==1.5.0
langchain-text-splitters==1.1.1

# Document loaders
pypdf==6.7.1
python-docx==1.2.0

# Voice
gTTS==2.5.4

# Utilities
python-dotenv==1.2.1
requests==2.32.5
